+++ src/library/framework_scan.cpp
@@ -428,7 +428,7 @@ void Scan(
    ParameterError parameterError;
    RunTimeError runTimeError;
    Error error;
-   #pragma omp parallel num_threads( nThreads )
+   #pragma omp parallel num_threads( static_cast< int >( nThreads ))
    try {
       dip::uint thread = static_cast< dip::uint >( omp_get_thread_num());
       std::vector< std::vector< uint8 >> buffers; // The outer one here is not a DimensionArray, because it won't delete() its contents
